requirements.txt
altgraph==0.17
astroid==2.4.2
auto-py-to-exe==2.7.8
autopep8==1.5.4
bottle-websocket==0.2.9
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
eel==0.12.4
future==0.18.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
moviepy==1.0.3
numpy==1.19.2
pefile==2019.4.18
Pillow==7.2.0
proglog==0.1.9
pycodestyle==2.6.0
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyinstaller-hooks-contrib==2020.9
pylint==2.6.0
pypiwin32==223
pytube3==9.6.4
pywin32==228
pywin32-ctypes==0.2.0
requests==2.24.0
six==1.13.0
toml==0.10.1
tqdm==4.49.0
typing-extensions==3.7.4.3
urllib3==1.25.10
whichcraft==0.6.1
wrapt==1.12.1
